services/analytics-service/requirements.txt
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1
apache-airflow-providers-amazon==8.13.0
apache-airflow-providers-google==10.10.1
kafka-python==2.0.2
pandas==2.1.4
pyarrow==14.0.1
pydantic==2.5.2
sqlalchemy==1.4.53
great-expectations==0.18.5
boto3==1.34.0
google-cloud-storage==2.10.0
psycopg2-binary==2.9.9
redis==5.0.1
structlog==23.2.0